Move createImportCharactersFolderIfNotExists to utilities.js
